chore: merge users and into user_management
