+++ backend/user_management/routing.py
@@ -1,4 +1,5 @@
 from django.urls import re_path
+
 from . import consumers
 
 websocket_urlpatterns = [
